Created a new test scene and created the trash spawner script
I made a new version of the white box level using Ethan's working scene as a base. It has the background, camera, player character, divers, and the new trash spawner script. The trash spawner takes the trash prefabs as inputs, then spawns and despawns them on a timer that can be adjusted in the inspector.